+++ aliensrematch/src/aliensrematch/bot2.java
@@ -703,8 +703,8 @@ public class bot2 {
 
 
 	// run the bot
-	int[] run() {
-		int[] ret = new int[2];
+	double[] run() {
+		double[] ret = new double[2];
 		int step = 0; // # of steps taken
