+++ aliensrematch/src/aliensrematch/bot3.java
@@ -664,8 +664,8 @@ public class bot3 {
 	}
 
 	// run the bot
-	int[] run() {
-		int[] ret = new int[2];
+	double[] run() {
+		double[] ret = new double[2];
 		int saved = 0; // # of crewmembers saved
 		int step = 0; // # of steps taken
